silence "conversion from double to float" warnings
Revert "silence "conversion from double to float" warnings"

This reverts commit b491b996e195232bfa4b70de8f6fdc3e34411f93.

silence "conversion from double to float" warnings

+++ src/common/menu/joystickmenu.cpp
@@ -140,7 +140,7 @@ DEFINE_ACTION_FUNCTION(IJoystickConfig, SetAxisResponseCurvePoint)
 	PARAM_INT(axis);
 	PARAM_INT(point);
 	PARAM_FLOAT(value);
-	self->SetAxisResponseCurvePoint(axis, point, value);
+	self->SetAxisResponseCurvePoint(axis, point, static_cast<float>(value));
 	return 0;
 }
